Sending and receiving simultaneously works!

+++ ipv8/src/main/java/nl/tudelft/ipv8/messaging/utp/channels/impl/receive/ConnectionIdTriplet.java
@@ -1,36 +1,16 @@
-/* Copyright 2013 Ivan Iljkic
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
 package nl.tudelft.ipv8.messaging.utp.channels.impl.receive;
 
 import nl.tudelft.ipv8.messaging.utp.channels.UtpSocketChannel;
 import nl.tudelft.ipv8.messaging.utp.channels.impl.UtpSocketChannelImpl;
 
-/**
- * DTO to put channel and the connection id's in one class.
- *
- * @author Ivan Iljkic (i.iljkic@gmail.com)
- */
 public class ConnectionIdTriplet {
 
     private UtpSocketChannel channel;
-    //incoming connection ID
-    private long incoming;
+    private short incoming;
     //outgoing is incoming connection ID + 1 as specified
-    private long outGoing;
+    private short outGoing;
 
-    public ConnectionIdTriplet(UtpSocketChannel channel, int incoming, long outgoing) {
+    public ConnectionIdTriplet(UtpSocketChannel channel, short incoming, short outgoing) {
         this.channel = channel;
         this.incoming = incoming;
         this.outGoing = outgoing;
@@ -68,22 +48,22 @@ public class ConnectionIdTriplet {
         return hash;
     }
 
-    public long getIncoming() {
+    public short getIncoming() {
         return incoming;
     }
 
 
-    public void setIncoming(long incoming) {
+    public void setIncoming(short incoming) {
         this.incoming = incoming;
     }
 
 
-    public long getOutGoing() {
+    public short getOutGoing() {
         return outGoing;
     }
 
 
-    public void setOutgoing(long outGoing) {
+    public void setOutgoing(short outGoing) {
         this.outGoing = outGoing;
     }
 }
